Создал новый класс хелпер, вынес туда проверки на валидацию
